+++ petJournal/petJournal/Commons/Enums/CustomFont.swift
@@ -1,0 +1,16 @@
+//
+//  FontType.swift
+//  petJournal
+//
+//  Created by Marcylene Barreto on 10/04/23.
+//
+
+import Foundation
+
+enum FontType: String {
+    case regular = "Regular"
+    case bold = "Bold"
+    case medium = "Medium"
+    case semiBold = "SemiBold"
+    case light = "Light"
+}
